BigQueryIoT/requirements.txt
google-cloud-bigquery==0.31.0
paho-mqtt==1.4.0